backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.9.1
certifi==2023.7.22
cffi==1.16.0
channels==4.0.0
charset-normalizer==3.3.0
click==8.1.7
constantly==23.10.4
coverage==7.3.2
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.8.0rc2
Django==4.2.5
django-filter==23.3
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
docutils==0.20.1
flake8==6.1.0
hyperlink==21.0.0
hypothesis==6.87.4
idna==3.4
incremental==22.10.0
Markdown==3.4.4
mccabe==0.7.0
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
sqlparse==0.4.4
Twisted==23.10.0
txaio==23.1.1
typing_extensions==4.8.0
urllib3==2.0.7
wheel==0.41.2
zope.interface==6.1
